[PC-901] Refactor: TabBarViewModel을 HomeViewModel에 통합
